+++ rust/blockstore/src/arrow/block/delta/mod.rs
@@ -1,4 +1,5 @@
 mod delta;
+pub(super) mod int32;
 mod storage;
 pub(super) mod string;
 pub(super) mod uint32;
